# docs/static/assets/openapi.json
{
  "components": {
    "schemas": {
...
      "RepositoryTO": {
        "type": "object",
        "required": [
          "assignedUsers",
          "description",
          "existingArtifacts",
          "id",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "existingArtifacts": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "assignedUsers": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "NewRepositoryTO": {
        "type": "object",
        "required": [
          "description",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ArtifactTO": {
        "type": "object",
        "required": [
          "description",
          "name",
          "repositoryId"
        ],
        "properties": {
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "svgPreview": {
            "type": "string"
          },
          "repositoryId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "updatedDate": {
            "format": "date-time",
            "type": "string"
          },
          "fileType": {
            "type": "string"
          }
        }
      },
...
      "UserUpdateTO": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string"
          },
          "username": {
            "minLength": 3,
            "type": "string",
            "maxLength": 50
          }
        }
      },
...
      "RepositoryUpdateTO": {
        "type": "object",
        "required": [
          "description",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UserInfoTO": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "minLength": 3,
            "type": "string",
            "maxLength": 50
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.1",
  "paths": {
...
    "/api/repo/{repositoryId}": {
      "get": {
        "summary": "Get a single Repository by providing its ID",
        "operationId": "getSingleRepository",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryTO"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Repository"
        ]
      },
      "delete": {
        "summary": "Delete a Repository if you own it",
        "operationId": "deleteRepository",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Repository"
        ]
      },
      "put": {
        "summary": "Update a Repository",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepositoryUpdateTO"
              }
            }
          },
          "required": true
        },
        "operationId": "updateRepository",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Repository"
        ]
      }
    },
    "/api/user/searchUsers/{typedName}": {
      "get": {
        "operationId": "searchUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserInfoTO"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "typedName",
            "required": true
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/user": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateTO"
              }
            }
          },
          "required": true
        },
        "operationId": "updateUser",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/repo": {
      "post": {
        "summary": "Create a new Repository",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewRepositoryTO"
              }
            }
          },
          "required": true
        },
        "operationId": "createRepository",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Repository"
        ]
      },
      "get": {
        "summary": "Get all Repositories",
        "operationId": "getAllRepositories",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryTO"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Repository"
        ]
      }
    },
...
    "/api/artifact/repository/{repositoryId}": {
      "get": {
        "operationId": "getArtifactsFromRepo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactTO"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "repositoryId",
            "required": true
          }
        ],
        "tags": [
          "Artifact"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  }
}